requirements.txt
matplotlib==3.7.5
numpy==1.24.3
numpy==1.24.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
Pillow==11.2.0
torch==2.3.0
torchvision==0.18.0
